two-step-optimization
In this update, first a full HMM with all the motifs is optimized.
Then, only motifs that are significantly enriched among positive
sequences and are still similar to their seeds are selected. A new HMM
with only these motifs is optimized (the emission probabilities remain
the same).

Also, the probability of each sequencing containing an instance of each
of the PFMs in the HMM is stored. This is then used by an R script to
construct a classifier for distinguishing positive from negative
sequences, and also to create a map of which ZFs are most likely to
bind to each positive sequence.